правки по замечаниям наставника
