+++ package.json
@@ -17,14 +17,13 @@
     "style-loader": "1.2.1",
     "webpack": "4.42.1",
     "webpack-cli": "3.3.11",
-    "webpack-dev-server": "^3.11.0"
+    "webpack-dev-server": "3.11.0"
   },
   "engines": {
     "node": "10"
   },
   "dependencies": {
     "flatpickr": "4.6.3",
-    "he": "1.2.0",
     "moment": "2.24.0"
   }
 }
