+++ src/components/edit-point/point-type-template.js
